Fix multi selection behaviour
The main issue here is that we had two (well, we actually had three, but one is v2) functions for sorting lists of files. The big exciting issue is that the two sorting algorithms handled capital letters differently 🤦‍♂️ . This PR bins the duplicate code and consolidating down to just one algorithm fixes the issue.



This also fixes an issue where in file mode, you couldn’t use the keyboard to go past the 99th file item due to it using the lazy loaded file list even though it didn’t do lazy loading.
